backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
requests==2.31.0
python-multipart==0.0.6
redis==5.0.1
celery==5.3.4
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4